Progress python script for network config.

+++ multinode-config/configure.py
@@ -11,21 +11,26 @@ import argparse
 from platform import system as system_name # Returns the system/OS name
 from os import system as system_call       # Execute a shell command
 from string import Template
+import json
 import paramiko
 from termcolor import colored, cprint
+
 
 def generate_network_config(template_filename, output_filename, replacement_dict):
     """
     Takes a template network configuration file and replaces the values.
     """
-    with open(template_filename, "r") as template_file:
-        data = template_file.read()
-        template = Template(data)
-        write_data = template.substitute(replacement_dict)
-        if not os.path.exists('./generated'):
-            os.makedirs('./generated')
-        with open(output_filename, "w+") as output_file:
-            output_file.write(write_data)
+    if not os.path.exists(template_filename):
+        cprint('Config file ' + template_filename + ' does not exist', 'red')
+    else:
+        with open(template_filename, "r") as template_file:
+            data = template_file.read()
+            template = Template(data)
+            write_data = template.substitute(replacement_dict)
+            if not os.path.exists('./generated'):
+                os.makedirs('./generated')
+            with open(output_filename, "w+") as output_file:
+                output_file.write(write_data)
 
 def ping(host):
     """
@@ -47,26 +52,76 @@ def test_ssh_connection(name, login_username, login_password):
     except:
         return False
 
+def load_json_config(filename):
+    """
+    Load configuration settings from a JSON file
+    """
+    with open(filename) as data_file:
+        data = json.load(data_file)
+        return data
+
+def validate_json(json_data, server_list):
+    """
+    Check that the JSON data is well formed.
+    """
+    for key in server_list:
+        if key not in json_data:
+            raise ValueError(key + ' value is not in the JSON file')
+
+def check_server(ipaddress, username, password, description):
+    """
+    Check a server via ping and SSH to see if it is alive. Return false if any check fails.
+    """
+    server_ok = True
+    if ping(ipaddress):
+        cprint(description + ' server contactable via ping.', 'green')
+    else:
+        cprint(description + ' server not contactable via ping.', 'red')
+        server_ok = False
+    if test_ssh_connection(ipaddress, username, password):
+        cprint(description + ' server contactable via SSH.', 'green')
+    else:
+        cprint(description + ' server not contactable via SSH.', 'red')
+        server_ok = False
+    return server_ok
+
 def main():
     """
-    Function docstring
+    Validate servers exist and change their network configuration
     """
     parser = argparse.ArgumentParser(description='Set up VMs for openstack-ansible deployment.')
-    parser.add_argument('--compute-address', type=str, default='', required=True,
-                        help='The address of the compute node to initially log in to.')
-    parser.add_argument('--compute-ip', type=str, default='', required=True,
-                        help='The static IP address of the compute node to assign to it.')
+    parser.add_argument('config_file', type=str, default='',
+                        help='A JSON file that contains the address for all the machines.')
     args = parser.parse_args()
-    if ping(args.compute_address):
-        cprint('Compute server contactable via ping.', 'green')
+    json_data = load_json_config(args.config_file)
+    # A list of all the servers we expect to be in the config file and ready for deployment
+    server_list = {'os_deploy_ip' : {'name':'os-deploy', 'filename':'deploy-network.conf'},
+                   'os_compute_ip' : {'name':'os-compute', 'filename':'compute-network.conf'},
+                   'os_network_ip' : {'name':'os-network', 'filename':'network-network.conf'},
+                   'os_storage_ip' : {'name':'os-storage', 'filename':'storage-network.conf'},
+                   'os_infrastructure_ip' : {'name':'os-infrastructure',
+                                             'filename':'infrastructure-network.conf'},
+                   'os_controller_ip' : {'name':'os-controller', 
+                                         'filename':'controller-network.conf'}}
+    validate_json(json_data, server_list)
+    default_username = 'vagrant'
+    default_password = 'vagrant'
+
+    all_servers_ok = True
+    for key in server_list:
+        if not check_server(json_data[key], default_username, default_password, server_list[key]['name']):
+            all_servers_ok = False
+
+    if not all_servers_ok:
+        cprint('At least one server is not available. Aborting.', 'red')
+        return
     else:
-        cprint('Compute server not contactable via ping.', 'red')
-    if test_ssh_connection(args.compute_address, 'vagrant', 'vagrant'):
-        cprint('Compute server contactable via SSH.', 'green')
-    else:
-        cprint('Compute server not contactable via SSH.', 'red')
-    replacement_dict = dict(primary_eth='eth0', dns1='8.8.8.8', dns2='8.8.4.4')
-    generate_network_config('infra-network.conf', './generated/infra-network.conf', replacement_dict)
+        cprint('All servers are contactable for deployment.', 'green')
+
+    json_data['primary_eth'] = 'eth0'
+    for key in server_list:
+        generate_network_config(server_list[key]['filename'], './generated/' +
+                                server_list[key]['filename'], json_data)
 
 if __name__ == '__main__':
     main()
